rename package to 'aws'
