+++ discovery/aws/ec2.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package amazon
+package aws
 
 import (
 	"context"
