+++ discovery/install/install.go
@@ -16,7 +16,7 @@
 package install
 
 import (
-	_ "github.com/prometheus/prometheus/discovery/amazon"       // register amazon
+	_ "github.com/prometheus/prometheus/discovery/aws"          // register aws
 	_ "github.com/prometheus/prometheus/discovery/azure"        // register azure
 	_ "github.com/prometheus/prometheus/discovery/consul"       // register consul
 	_ "github.com/prometheus/prometheus/discovery/digitalocean" // register digitalocean
